% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectBrainCell.R
\docType{class}
\name{TrenaProjectBrainCell-class}
\alias{TrenaProjectBrainCell-class}
\alias{.TrenaProjectBrainCell}
\alias{TrenaProjectBrainCell}
\title{TrenaProjectBrainCell-class}
\usage{
TrenaProjectBrainCell(quiet = TRUE)
}
\value{
An object of the TrenaProjectBrainCell class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
